Devide store to modules
