+++ client/store/flash/index.js
@@ -1,0 +1,8 @@
+export default {
+  state: {
+    messages: [],
+  },
+  getters: {},
+  actions: {},
+  mutations: {},
+}
